+++ Assets/ImageSource.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 576789de97f45734a8a99b3116e2391e
-PrefabImporter:
+guid: bd7b4703a4195604b840282a38ba6d6f
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
