+++ Assets/LoginManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4daaa3ef19a10804c965bdc0e217e1ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
