+++ Assets/Material.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 576789de97f45734a8a99b3116e2391e
-PrefabImporter:
+guid: f49c5af05cf399c4d9b8d229642fe9d8
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
