+++ Assets/Scenes/LoginScene.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 576789de97f45734a8a99b3116e2391e
-PrefabImporter:
+guid: e3a384f64403ba64680bb8b7a542f7f2
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
