+++ Assets/ToonyTinyPeople.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 576789de97f45734a8a99b3116e2391e
-PrefabImporter:
+guid: 84ca10afaeef50a4c8223da1b7a727a5
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
