+++ Assets/ToonyTinyPeople/TT_ww1_demo.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 576789de97f45734a8a99b3116e2391e
-PrefabImporter:
+guid: ae3ddc1ae9126bd4daf24f6c11d7b1d9
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
