+++ Assets/ToonyTinyPeople/TT_ww1_demo/animation.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: daab072fddaf8504da4456fcdbf3f9b3
+folderAsset: yes
+timeCreated: 1522685496
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
