+++ Assets/ToonyTinyPeople/TT_ww1_demo/animation/m_weapon_death_A.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cdae11d1196d7c9f245dbd967eebd5cc2123bb8093b3bbf733b6ac0796c8290a
+size 259728
