+++ Assets/ToonyTinyPeople/TT_ww1_demo/animation/m_weapon_idle_A.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9e96413880390cc55172649bde14a714411609fc28acc7c936ce89dcd6f4b074
+size 387600
