+++ Assets/ToonyTinyPeople/TT_ww1_demo/animation/m_weapon_run.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:13cf3be6f5417870538444880012453c33077eb93028a84f2a585a15c77bd362
+size 180336
