+++ Assets/ToonyTinyPeople/TT_ww1_demo/animation/m_weapon_shoot.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:43f1a2d4428891cd7369ec727ab6705c2e32500fe97118bb60aa5becec0bdc7d
+size 202944
