+++ Assets/ToonyTinyPeople/TT_ww1_demo/models.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 3c7639bc4b59b114db36d210e93e81f6
+folderAsset: yes
+timeCreated: 1522685496
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
