+++ Assets/ToonyTinyPeople/TT_ww1_demo/models/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 2281fab91cd1e824c994e86c67d95040
+folderAsset: yes
+timeCreated: 1522685496
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
