+++ Assets/ToonyTinyPeople/TT_ww1_demo/models/Materials/TT_WW1_demo_material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fc99f4b4a172f1e499845706d0122d79
+timeCreated: 1522687106
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
