+++ Assets/ToonyTinyPeople/TT_ww1_demo/models/Materials/TT_ww1_demo_texture.tga
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:63aa60aea251e68204eec75968bf32b142c98ab9ec2449eacfbcb1d0ab0d24f6
+size 29500
