+++ Assets/ToonyTinyPeople/TT_ww1_demo/models/TT_ww1_demo_british_A.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b15c46ede81ff4c98908030d561c72fe7c35a3c0bb11ec1ba0782ee3624adbdf
+size 93536
