+++ Assets/ToonyTinyPeople/TT_ww1_demo/models/TT_ww1_demo_german_A.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4ce8315c749c61c4a427000b5631d23a819116479a4c4b71460d018c3db5debe
+size 97392
