+++ Assets/ToonyTinyPeople/TT_ww1_demo/models/TT_ww1_demo_german_B.FBX
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1cb7021fbad191882334557e4fb607acdf622bcaced0745d4278041431703489
+size 112880
