+++ Assets/ToonyTinyPeople/TT_ww1_demo/prefabs.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f264126355ada3349a5709d5f55f1c0e
+folderAsset: yes
+timeCreated: 1522685496
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
