+++ Assets/ToonyTinyPeople/TT_ww1_demo/prefabs/TT_ww1_demo_british_A.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa05473e4d04e834eb7a29b53251c8cf
+timeCreated: 1522688031
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
