+++ Assets/ToonyTinyPeople/TT_ww1_demo/prefabs/TT_ww1_demo_german_A.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 789396300d039cc4fa674d36eddb10a7
+timeCreated: 1522688035
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
