+++ Assets/ToonyTinyPeople/TT_ww1_demo/sample_scene.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 04848fb328ab2e049a3f8d774f55a70f
+folderAsset: yes
+timeCreated: 1522685496
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
