+++ Assets/ToonyTinyPeople/TT_ww1_demo/sample_scene/TT_ww1_sample_scene.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb4fed3485d830b49810b12950c81666
+timeCreated: 1522685497
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
