+++ Assets/ToonyTinyPeople/TT_ww1_demo/sample_scene/TT_ww1_sample_sceneSettings.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea308c0fdab469e41990615db435ed4f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
