+++ Assets/ToonyTinyPeople/TT_ww1_demo/sample_scene/animation_sample.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 43be405b123e55543ad3df22200462ec
+timeCreated: 1524584045
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
