+++ Assets/ToonyTinyPeople/TT_ww1_demo/sample_scene/quad.tga
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cc2752cf773a102d8dc086c202b6ded1cd850796f06c07846a6e86c2028ac974
+size 32652
